CTCTRADERS-687: move to package
