+++ app/utils/TransportDetailsCheckYourAnswersHelper.scala
@@ -24,6 +24,22 @@ import uk.gov.hmrc.viewmodels._
 
 class TransportDetailsCheckYourAnswersHelper(userAnswers: UserAnswers) {
 
+
+  def inlandMode: Option[Row] = userAnswers.get(InlandModePage) map {
+    answer =>
+      Row(
+        key     = Key(msg"inlandMode.checkYourAnswersLabel", classes = Seq("govuk-!-width-one-half")),
+        value   = Value(lit"$answer"),
+        actions = List(
+          Action(
+            content            = msg"site.edit",
+            href               = routes.InlandModeController.onPageLoad(lrn, CheckMode).url,
+            visuallyHiddenText = Some(msg"site.edit.hidden".withArgs(msg"inlandMode.checkYourAnswersLabel"))
+          )
+        )
+      )
+  }
+
   def idCrossingBorder: Option[Row] = userAnswers.get(IdCrossingBorderPage) map {
     answer =>
       Row(
