+++ conf/app.routes
@@ -181,7 +181,7 @@ POST       /:lrn/transport-details/add-id-at-departure                        co
 GET        /:lrn/transport-details/change-add-id-at-departure                controllers.transportDetails.AddIdAtDepartureController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
 POST       /:lrn/transport-details/change-add-id-at-departure                controllers.transportDetails.AddIdAtDepartureController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
 
-GET        /:lrn/transport-details/inland-mode                        controllers.InlandModeController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-POST       /:lrn/transport-details/inland-mode                          controllers.InlandModeController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-GET        /:lrn/change-transport-details/inland-mode                    controllers.InlandModeController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
-POST       /:lrn/change-transport-details/inland-mode                    controllers.InlandModeController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
+GET        /:lrn/transport-details/inland-mode                        controllers.transportDetails.InlandModeController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
+POST       /:lrn/transport-details/inland-mode                          controllers.transportDetails.InlandModeController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
+GET        /:lrn/change-transport-details/inland-mode                    controllers.transportDetails.InlandModeController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
+POST       /:lrn/change-transport-details/inland-mode                    controllers.transportDetails.InlandModeController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
